+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -7,11 +7,11 @@ mod list;
 mod mkdir;
 mod mv;
 mod print;
-mod read;
+mod read_text;
 mod rm;
 mod rmdir;
 mod touch;
-mod write;
+mod write_text;
 
 use crate::utils::pckg;
 use duckscript::types::command::Commands;
@@ -31,11 +31,11 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(mkdir::create(&package))?;
     commands.set(mv::create(&package))?;
     commands.set(print::create(&package))?;
-    commands.set(read::create(&package))?;
+    commands.set(read_text::create(&package))?;
     commands.set(rm::create(&package))?;
     commands.set(rmdir::create(&package))?;
     commands.set(touch::create(&package))?;
-    commands.set(write::create(&package))?;
+    commands.set(write_text::create(&package))?;
 
     Ok(())
 }