+++ duckscript_sdk/src/sdk/std/fs/read_text/mod.rs
@@ -12,11 +12,11 @@ pub(crate) struct CommandImpl {
 
 impl Command for CommandImpl {
     fn name(&self) -> String {
-        pckg::concat(&self.package, "Write")
+        pckg::concat(&self.package, "ReadText")
     }
 
     fn aliases(&self) -> Vec<String> {
-        vec!["writefile".to_string()]
+        vec!["readfile".to_string(), "read_text_file".to_string()]
     }
 
     fn help(&self) -> String {
@@ -29,15 +29,13 @@ impl Command for CommandImpl {
 
     fn run(&self, arguments: Vec<String>) -> CommandResult {
         if arguments.is_empty() {
-            CommandResult::Error("File name and text not provided.".to_string())
-        } else if arguments.len() == 1 {
-            CommandResult::Error("Text not provided.".to_string())
+            CommandResult::Error("File name not provided.".to_string())
         } else {
-            let result = io::write_text_file(&arguments[0], &arguments[1]);
+            let result = io::read_text_file(&arguments[0]);
 
             match result {
-                Ok(_) => CommandResult::Continue(Some("true".to_string())),
-                Err(_) => CommandResult::Continue(Some("false".to_string())),
+                Ok(text) => CommandResult::Continue(Some(text)),
+                Err(error) => CommandResult::Error(error.to_string()),
             }
         }
     }